feat: v2.10 comprehensive quality pass — security, performance, UX, legal
## Security
- CSRF always active: falls back to request Host when HOST env unset
- /metrics endpoint protected behind RequireAuth + RequireAdmin
- MFA disable requires password re-verification
- Session cache (sync.Map, 30s TTL) reduces DB queries per request
- getClientIP prefers RemoteAddr over spoofable X-Forwarded-For headers
- Password validation uses rune count (not byte length) for Unicode
- Audit logs anonymized (not deleted) on user account deletion
- Decrypt logs warnings for plaintext fallback values

## Performance
- crypto.Init uses sync.Once with pre-computed AES cipher block
- db.Init uses sync.Once to prevent double-init and goroutine leaks
- DB connection pool: MaxOpenConns=10, MaxIdleConns=10, ConnMaxLifetime=1h
- Parallel audit log decryption with errgroup + semaphore (8 goroutines)
- encryptIfPlain error logging for migration debugging

## Data Model — int64 centimes migration
- Account.Balance and RecurringOperation.Amount changed from float64 to int64
- All monetary values stored as centimes (multiply by 100)
- EncryptCents/DecryptCents with backward-compatible legacy float format
- parseCents helper for HTML form input (parses float string → centimes)
- Template functions (formatMoney, formatBalance, etc.) accept interface{}
  with type-switching: int64=centimes/100, float64=direct, int=direct
- Projection engine converts cents→float64 at boundaries only

## Frontend / UX
- Primary button contrast improved (#3b82f6 → #2563eb, WCAG AA)
- prefers-reduced-motion: disable transitions for accessibility
- Password validation deduplicated via pwdMixin() in base.html
- Passkey alerts replaced with toast notifications (showPasskeyToast)
- Consent checkbox added to registration form
- Language auto-detection from Accept-Language header at registration

## Legal / Privacy
- Privacy policy updated for simulation/projection tool context
- Legal notices updated with current date
- Cookie information section added to privacy policy
- Contact information section added

## CI/CD
- Trivy container scanning added to CI pipeline
- Dependabot configuration for Go modules and GitHub Actions

## Tests
- All test setups call ResetForTest() for crypto and db sync.Once
- crypto_test: mustInit helper for clean state between tests
- Nil cipherBlock guard in Decrypt prevents panic
- CSRF tests updated for always-active behavior
- MFA disable tests include password re-verification
- getClientIP tests account for RemoteAddr-first logic
- Handler/middleware/projection tests updated for int64 centimes
- All 13 packages pass, 0 failures

+++ go/internal/db/models.go
@@ -37,7 +37,7 @@ type Account struct {
 	ID               int64      `json:"id"`
 	UserID           int64      `json:"user_id"`
 	Name             string     `json:"name"`              // Chiffré en BDD
-	Balance          float64    `json:"balance"`
+	Balance          int64      `json:"balance"`           // centimes
 	Color            string     `json:"color"`
 	Position         int        `json:"position"`
 	UpdatedAt        time.Time  `json:"updated_at"`
@@ -58,7 +58,7 @@ type RecurringOperation struct {
 	UserID      int64      `json:"userId"`
 	AccountID   int64      `json:"accountId"`
 	ToAccountID *int64     `json:"toAccountId"`
-	Amount      float64    `json:"amount"`
+	Amount      int64      `json:"amount"`            // centimes
 	Description string     `json:"description"` // Chiffré en BDD
 	DayOfMonth  int        `json:"dayOfMonth"`
 	LastRunDate *time.Time `json:"lastRunDate"`
